fix: 모킹 및 디자인 세부 수정
